alle Bit-oriented Befehle implementiert. Goto Befehl auch(dafür logik mit pcl und pclath in Storage.cs verbessert. Abarbeitung der Befehle über eine While Schleife. Es gibt aber noch Probleme mit der Abbruchbediengung und der Anzeige des aktuellen Befehls

+++ PicSimulator/PicSimulator/Simulator/Flags.cs
@@ -15,7 +15,7 @@ namespace PicSimulator.Simulator
         }
 
         #region Statusregister Flags
-        private static byte statusRegister = Storage.dataMemory[3];
+        private static int statusRegister = Storage.dataMemory[3];
 
         /// <summary>
         /// set or delete Zeroflag in the statusregister
